+++ src/components/Milestone/Milestone.tsx
@@ -3,125 +3,38 @@ import './Milestone.css'
 import hljs from 'highlight.js'
 import parse from 'html-react-parser'
 import { ArrowLeftCircle, ArrowRightCircle, Calendar } from 'iconoir-react'
-import { DateTime } from 'luxon'
-import { useEffect, useMemo, useRef, useState } from 'react'
+import { useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useQueryParams, withDefault } from 'use-query-params'
 
-import {
-  FilterByQueryparam,
-  OrderByQueryParam,
-  StatusSuccess,
-} from '../../constants/globalConstants'
-import { validateForm } from '../../dashboard/utils/helpers/schema'
-import { useGetJSON } from '../../logic/api/fetchData'
+import { FilterByQueryparam, OrderByQueryParam } from '../../constants/globalConstants'
 import { asEnumParam, asRegexArrayParam } from '../../logic/params'
 import Facets from '../Facets/Facets'
 import OrderByDropdown from '../OrderByDropdown'
 import MonthCard from './Card/MonthCard'
+import useMilestoneFetch from './fetch'
 import { getMonths } from './helper'
-import { milestoneSchema } from './schema'
 
 const Milestone = () => {
   const { t } = useTranslation()
 
-  const [selectedIndices, setSelectedIndices] = useState<any>(null)
-  const [facetsResetKey, setFacetsResetKey] = useState(0)
-  const [cursor, setCursor] = useState(0)
+  const {
+    parsedTimeline: timeline,
+    timelineError,
+    errorArticles,
+    errorGithub,
+  } = useMilestoneFetch()
 
   const containerRef = useRef<HTMLDivElement>(null)
+  const [selectedIndices, setSelectedIndices] = useState<any>(null)
+  const [facetsResetKey, setFacetsResetKey] = useState(0)
   const [isDragging, setIsDragging] = useState(false)
   const [startX, setStartX] = useState(0)
   const [scrollLeft, setScrollLeft] = useState(0)
   const MONTH = 12
 
-  const {
-    data: dataGithub,
-    error: errorGithub,
-    status: statusGithub,
-  } = useGetJSON({
-    url: import.meta.env.VITE_WIKI_EVENTS,
-  })
-
-  const {
-    data: articles,
-    error: errorArticles,
-    status: statusArticles,
-  } = useGetJSON({
-    url: '/api/articles?limit=500',
-    delay: 0,
-  })
-
-  const { parsedTimeline, timelineError } = useMemo(() => {
-    if (!dataGithub) return {}
-    let json = {}
-
-    //Validate Github Data
-    try {
-      if (statusGithub !== StatusSuccess) {
-        return { parsedTimeline: null, timelineError: null }
-      }
-
-      if (!dataGithub || typeof dataGithub !== 'string') {
-        return {
-          parsedTimeline: null,
-          timelineError: t('milestone.error.emptyData'),
-        }
-      }
-
-      if (statusGithub === StatusSuccess) {
-        json = JSON.parse(dataGithub.replace(/^```json\n/, '').replace(/\n```$/, ''))
-      }
-    } catch (e) {
-      return {
-        parsedTimeline: null,
-        timelineError: t('milestone.error.notValid'),
-      }
-    }
-
-    const { valid, errors } = validateForm(json, milestoneSchema)
-
-    if (!valid) {
-      const detail = errors?.[0]?.message ?? t('milestone.error.structure')
-      return {
-        parsedTimeline: null,
-        timelineError: `${t('milestone.error.malformed')} ${detail}`,
-      }
-    }
-
-    // Merging data from API and Github
-    const articlesByYear = (articles?.results ?? []).reduce((acc, article) => {
-      const year = DateTime.fromISO(article.publication_date).year
-      const issue = article.issue.pid.replace(/jdh0+(\d+)/, (m, n) => n)
-      const title = article.data.title[0].replace('# ', '')
-
-      if (!acc[year]) {
-        acc[year] = []
-      }
-
-      acc[year].push({
-        date: article.publication_date,
-        title: title,
-        issue: issue,
-        pid: article.abstract.pid,
-      })
-
-      return acc
-    }, {})
-
-    const articlesAndGithubData = Object.keys(json).reduce((acc, year) => {
-      acc[year] = {
-        ...json[year],
-        articles: articlesByYear[year] ?? [],
-      }
-      return acc
-    }, {})
-
-    return { parsedTimeline: articlesAndGithubData, timelineError: null }
-  }, [articles, dataGithub])
-
   // Set up years for dropdown
-  const years = Object.keys(parsedTimeline ?? [])
+  const years = Object.keys(timeline ?? [])
     .map((year) => ({ value: year, label: year }))
     .reverse()
 
@@ -133,23 +46,22 @@ const Milestone = () => {
     [FilterByQueryparam]: asRegexArrayParam(),
   })
 
-  const months = getMonths(MONTH, orderByYear, cursor)
+  const months = getMonths(timeline, MONTH)
 
-  const currentYearData = parsedTimeline?.[orderByYear] ?? {
-    articles: [] as any[],
-    callForPapers: [] as any[],
-    conferences: [] as any[],
-    releases: [] as any[],
-  }
+  const allEvents = useMemo(() => {
+    if (!timeline) return []
 
-  const milestoneItems = currentYearData
-    ? [
-        ...currentYearData.articles.map((item) => ({ ...item, type: 'Articles' })),
-        ...currentYearData.callForPapers.map((item) => ({ ...item, type: 'Call for Papers' })),
-        ...currentYearData.conferences.map((item) => ({ ...item, type: 'Conferences' })),
-        ...currentYearData.releases.map((item) => ({ ...item, type: 'Releases' })),
+    return Object.keys(timeline).flatMap((year) => {
+      const eventsOneYear = timeline[year]
+
+      return [
+        ...eventsOneYear.articles.map((item) => ({ ...item, type: 'Articles' })),
+        ...eventsOneYear.callForPapers.map((item) => ({ ...item, type: 'Call for Papers' })),
+        ...eventsOneYear.conferences.map((item) => ({ ...item, type: 'Conferences' })),
+        ...eventsOneYear.releases.map((item) => ({ ...item, type: 'Releases' })),
       ]
-    : []
+    })
+  }, [timeline])
 
   const milestoneDimensions = [
     {
@@ -163,28 +75,12 @@ const Milestone = () => {
 
   const visibleItems =
     selectedIndices === null
-      ? milestoneItems
-      : milestoneItems.filter((_, index) => selectedIndices?.includes(index))
+      ? allEvents
+      : allEvents.filter((_, index) => selectedIndices?.includes(index))
 
   const handleFacetAll = () => {
     setSelectedIndices(null)
     setFacetsResetKey((k) => k + 1)
-  }
-
-  useEffect(() => {
-    setCursor(0)
-  }, [orderByYear])
-
-  if (timelineError || errorArticles || errorGithub || parsedTimeline === undefined) {
-    const err = hljs.highlight(
-      'typescript',
-      `${t('milestone.error.general')} ${timelineError || errorArticles || errorGithub}`,
-    )
-    return (
-      <pre className="hljs" data-test="error-message">
-        <div>{parse(err.value)}</div>
-      </pre>
-    )
   }
 
   // --- Scrolling with arrows ---
@@ -224,6 +120,18 @@ const Milestone = () => {
     container.scrollLeft = scrollLeft - walk
   }
 
+  if (timelineError || errorArticles || errorGithub || timeline === undefined) {
+    const err = hljs.highlight(
+      'typescript',
+      `${t('milestone.error.general')} ${timelineError || errorArticles || errorGithub}`,
+    )
+    return (
+      <pre className="hljs" data-test="error-message">
+        <div>{parse(err.value)}</div>
+      </pre>
+    )
+  }
+
   return (
     <div className="milestone-wrapper">
       <p>Key dates and events for {orderByYear}</p>
@@ -241,7 +149,7 @@ const Milestone = () => {
             <Facets
               key={`${orderByYear}-${facetsResetKey}`}
               dimensions={milestoneDimensions}
-              items={milestoneItems}
+              items={allEvents}
               onSelect={(_, indices) => setSelectedIndices(indices)}
               className="milestone"
             />
